requirements.txt
channels==3.0.3
channels-redis==3.2.0
Django==3.2.4
django-allauth==0.44.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
Pillow==8.2.0
